Fix StyleCop type and filename warnings
