+++ Jellyfin.Plugin.Newsletters/Clients/Discord/DiscordWebhook.cs
