+++ Jellyfin.Plugin.Newsletters/Clients/Email/SmtpMailer.cs
